Move app/importers to lib/importers
This stops the code from being autoloader by the 
rails application. Additionally by moving it to lib
we can ensure clean separation thus making it
easier to extract the importer code into a gem in
the future.
