--- app/importers/daily_calls/daily_calls.rb
@@ -1,3 +1,0 @@
-module DailyCalls
-  autoload :Twilio, 'daily_calls/twilio'
-end
