+++ lib/importers.rb
@@ -1,3 +1,4 @@
 module Importers
+  autoload :DailyCalls, 'importers/daily_calls'
   autoload :WhereDidYouHearAboutUs, 'importers/where_did_you_hear_about_us'
 end
