+++ lib/importers/daily_calls.rb
@@ -1,0 +1,5 @@
+module Importers
+  module DailyCalls
+    autoload :Twilio, 'importers/daily_calls/twilio'
+  end
+end
