+++ spec/features/import_google_where_did_you_hear_data_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
-require_relative '../../lib/importers'
+require 'importers'
 
 RSpec.feature 'Importing google `where did you hear about us` data' do
   let(:csv) do
