+++ spec/importers/daily_calls/twilio/saver_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
+require 'importers'
 
-RSpec.describe DailyCalls::Twilio::Saver do
+RSpec.describe Importers::DailyCalls::Twilio::Saver do
   subject { described_class.new(calls: calls) }
 
   describe '.valid_calls_by_date' do
